simplify: merge web_taint.rs into context.rs (#731)
Both modules handle context injection for the model — context.rs builds
security context with fingerprinting/deduplication, web_taint.rs detects
web taint and generates warnings. They belong together conceptually.

This reduces nucleus-claude-hook from 18 to 17 modules with no behavior
change. All 173 tests pass.
